Splitting up /src/lib/api.ts
Created files for each function, added JSDOC and included JSDOC types as Typescript Types when possible.
