+++ src/pages/SubscriptionConfirmation.tsx
@@ -1,6 +1,7 @@
 import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router';
-import { syncStripeData } from '../lib/api';
+
+import {syncStripeData} from "../api/syncStripeData";
 
 const SubscriptionConfirmation: React.FC = () => {
   const navigate = useNavigate();
