+++ src/sync/versions.ts
@@ -1,8 +1,9 @@
 import { jwtDecode } from 'jwt-decode'
-import { pullUpdatesForVersions, pushUpdatesForVersions } from '../lib/api'
 import { AuthStorage, SyncStorage } from '../lib/storage'
 import { db } from '../storage'
 import { VersionedResource } from '../storage/schemas/versionedResource'
+import {pullUpdatesForVersions} from "../api/pullUpdatesForVersions";
+import {pushUpdatesForVersions} from "../api/pushUpdatesForVersions";
 
 const CHECKPOINT = 'version-checkpoint'
